[ConsulHealthCheck] Allow to filter on service names
Before this patch, one had to know exactly the ids of the service to
check. Now we can simply configure to check all checks from a given list
of services.
This will allow to easily remove services without fearing for ids not
being the same between registration having happened before or after the
service removal in check.

Change-Id: I4fad6491976ff940bbc3b1f94add6837a402eba8