Adapt to changes from ruby 2.7+
Starting with ruby 2.7, usage of Proc.new without a body is discouraged
(warning and then error).
This patch makes sure we are ready for this.

This part of the choregraphie code is a bit arcane for non-initiated and
I had trouble myself understanding the full picture 5 years after
writing it so here is my understanding.

Rationale:
- most people writing choregraphies are not heavy ruby users, quite
  often not even experienced chef users. They consider ruby/chef to be
  some kinf of magic
- specifically rules of bindings (what kind of definitions can be called
  within a block, depending of where it was defined) are not very clear.
  Usage of choregraphie dsl within chef dsl does not make any of this
  clearer
- a reasonable expectation as a user is to be able to call methods from
  within choregraphie (in the core block or in a before/after block)
  that were defined anywhere in the same file
- another reasonable expectation is to be able to call the same
  methods/variables inside and outside choregraphie block

Lines touched by this patch are doing exactly this: they store the
context in which choregraphie object was initialized and refer any
missing reference (method) to this context.

Fixes #62.

Change-Id: I5f4f9c8ba1dcf67e0e1bae1ba3fa17d29cea385a