Check several times maintenance status
We've observed in rare cases that consul maintenance status was not
correct shortly after a restart (and for ~5s).
It seems to happen only on our setup and can be reproduced with:

```

set -o pipefail

while true; do
  curl -f -s localhost:8500/v1/agent/checks | jq ._node_maintenance.Notes -r > /tmp/res
  code=$?
  if [[ "$code" -eq 0 ]]; then
    if grep -q test /tmp/res; then
      # nothing
      echo -n .
    else
      echo "$(date) reproduced!"
    fi
  else
    echo "$(date) error in curl"
  fi

done
```

and doing a `sudo systemctl restart consul`

In front the difficult to reproduce, we won't open a ticket upstream for
now but critically need a safeguard since it can lead to incidents. For
instance an incident that really happen is the release of the
consul_lock primitve while keeping consul maintenance status enabled. A
whole cluster collapsed because of this.

Change-Id: Ia8c052d5f082e4ffcbe70360064feb0de17e00c9